TD and FTA Azurefunction Updates
